git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@4169 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ asl/xml/Nodes.cpp
@@ -2404,7 +2404,7 @@ enum {
 };
 
 void dumpType(unsigned short theType) {
-    std::cerr << "Type = " << (void*)(int)theType << ", "<< dom::NodeTypeName[theType&0xf] << ",";
+    std::cerr << "Type = " << (void*)(ptrdiff_t)theType << ", "<< dom::NodeTypeName[theType&0xf] << ",";
     if (theType&hasName) cerr << "N ";
     if (theType&hasValue) cerr << "V ";
     if (theType&hasTypedValue) cerr << "TV ";
